+++ themes/GamePlatform_VelzonSaas/assets/js/games/application/components/game-boards/backgammon-container/backgammon-container.component.ts
@@ -508,7 +508,7 @@ export class BackgammonContainerComponent implements OnInit, OnDestroy, AfterVie
         this.height = Math.min( window.innerHeight - 40, this.width * 0.6 );
         
         const buttons = this.backgammonBoardButtons?.nativeElement as HTMLElement;
-        const btnsOffset = 15; //Cheating. Could not get the height.
+        const btnsOffset = 68; //Cheating. Could not get the height.
         if ( buttons ) {
             buttons.style.top = `${this.height / 2 + btnsOffset}px`;
             buttons.style.right = `${this.width * 0.12}px`;
@@ -518,13 +518,13 @@ export class BackgammonContainerComponent implements OnInit, OnDestroy, AfterVie
         if ( dices ) {
             // Puts the dices on right side if its my turn.
             if ( this.myTurn() ) {
-                dices.style.left = `${this.width / 2 + 20}px`;
+                dices.style.left = `${this.width / 2 - 95}px`;
                 dices.style.right = '';
             } else {
-                dices.style.right = `${this.width / 2 + 20}px`;
+                dices.style.right = `${this.width / 2 - 95}px`;
                 dices.style.left = '';
             }
-            dices.style.top = `${this.height / 2 - btnsOffset}px`;
+            dices.style.top = `${this.height / 2 + btnsOffset}px`;
         }
     }