+++ themes/GamePlatform_VelzonSaas/assets/js/games/application/components/side-bars/side-bars.module.ts
@@ -10,15 +10,13 @@ import { GameChatComponent } from './game-chat/game-chat.component';
 import { GameRoomsComponent } from './game-rooms/game-rooms.component';
 import { GamePlayersComponent } from './game-players/game-players.component';
 import { GameStatisticsComponent } from './game-statistics/game-statistics.component';
-import { BackgammonVariantsComponent } from './backgammon-variants/backgammon-variants.component';
 
 @NgModule({
     declarations: [
         GameChatComponent,
         GameRoomsComponent,
         GamePlayersComponent,
-        GameStatisticsComponent,
-        BackgammonVariantsComponent
+        GameStatisticsComponent
     ],
     imports: [
         CommonModule,
@@ -32,8 +30,7 @@ import { BackgammonVariantsComponent } from './backgammon-variants/backgammon-va
         GameChatComponent,
         GameRoomsComponent,
         GamePlayersComponent,
-        GameStatisticsComponent,
-        BackgammonVariantsComponent
+        GameStatisticsComponent
     ]
 })
 export class SideBarsModule { }